+++ drizzle/migrations/0006_freezing_jack_murdock.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `tasks` ADD `associated_backup_policy_id` integer;