+++ src/components/dashboard-header.tsx
@@ -2,6 +2,7 @@
 import { LogoutButton } from "@app/login/components";
 import { Button } from "@comp/button";
 import { Dialog, DialogContent, DialogTitle } from "@comp/dialog";
+import { Separator } from "@comp/separator";
 import { cn } from "@lib/utils";
 import Image from "next/image";
 import { useState } from "react";
@@ -19,7 +20,7 @@ export function DashboardHeader () {
                     <Image src="/logo.svg" alt="MongoDB Logo" width={32} height={32} />
                     <div className="text-xl font-bold hidden sm:inline-flex">Mongo Backup Service</div>
                     <Button onClick={() => setIsChangeLogOpen(true)} variant={"ghost"} className="text-muted-foreground hidden sm:inline-flex">
-                        v0.1.1
+                        v0.1.2
                     </Button>
                 </div>
                 <LogoutButton />
@@ -30,10 +31,16 @@ export function DashboardHeader () {
                     <div className={cn([
                         "[&_h1]:text-xl [&_h2]:text-lg [&_h3]:text-base [&_h4]:text-sm [&_h5]:text-xs [&_h6]:text-xs",
                         "[&_h1]:font-bold [&_h2]:font-bold [&_h3]:font-bold [&_h4]:font-bold [&_h5]:font-bold [&_h6]:font-bold",
-                        "[&_ul]:list-disc [&_ol]:list-decimal [&_li]:ml-4 [&_li]:mt-2",
+                        "[&_ul]:list-disc [&_ol]:list-decimal [&_li]:ml-4 [&_li]:mt-2 [&_ul]:mb-4",
                         "flex flex-col gap-2 p-4"
                     ])}>
-                        <ReactMarkdown children={changelogMdFile.default} />
+                        <ReactMarkdown
+                            components={{
+                                hr: () => (<Separator className="my-4" />)
+                            }}
+                        >
+                            {changelogMdFile.default}
+                        </ReactMarkdown>
                     </div>
                 </DialogContent>
             </Dialog>