+++ CHANGELOG.md
@@ -1,7 +1,18 @@
+## 0.1.2
+
+- Fixed: Authentication issue
+- Fixed: 'Clear All' tasks button not disassociating deleted task with policy/database.
+- Fixed: Now tasks remember the policy they were associated with, if any, instead of only the latest one.
+- Changed: 'Backup Delete' operation is treated as a task now
+
+---
+
 ## 0.1.1
 
-- Fixed bug where expired backups were not being deleted properly
-- Added this changelog dialog
+- Fixed: Expired backups were not being deleted properly
+- Added: Changelog dialog
+
+---
 
 ## 0.1.0